moves the BasicRoutingInterface to it's own file, to avoid circular dependency warnings when using tools like rollup
