+++ lib/route-tree-node.js
@@ -21,7 +21,7 @@
 import page from 'page';
 const Context = page.Context;
 import RouteData from './route-data.js';
-import routingMixin from '../routing-mixin.js';
+import { BasicRoutingInterface } from '../routing-interface.js';
 
 class RouteTreeNode {
   /** @param {!RouteData} data */
@@ -297,7 +297,7 @@ class RouteTreeNode {
 
         return nextEntry();
       } else if (currentExitNode.getValue().element) {
-        const routingElem = /** @type {!routingMixin.Type} */ (currentExitNode.getValue().element);
+        const routingElem = /** @type {!BasicRoutingInterface} */ (currentExitNode.getValue().element);
         if (!routingElem.routeExit) {
           throw new Error(`Element '${currentExitNode.getValue().tagName}' does not implement routeExit`);
         }
@@ -319,7 +319,7 @@ class RouteTreeNode {
       if (!currentEntryNode) {
         return;
       } else if (currentEntryNode.getValue().element) {
-        const routingElem = /** @type {!routingMixin.Type} */ (currentEntryNode.getValue().element);
+        const routingElem = /** @type {!BasicRoutingInterface} */ (currentEntryNode.getValue().element);
         if (!routingElem.routeEnter) {
           throw new Error(`Element '${currentEntryNode.getValue().tagName}' does not implement routeEnter`);
         }
